Extend helpers/derivedFeatures.ts with core feature constructors
Move createDerivedFeature, previewOffsetFeatures, joinOpenProfiles, and
clearStaleConstraints from projectStore.ts into helpers/derivedFeatures.ts.
The previewOffsetFeatures wrapper now calls its internal counterpart
directly. Consumers now import these from the helper module.
