+++ src/store/INDEX.md
@@ -30,7 +30,7 @@ Zustand store. The single source of truth for the current `.camj` project. **All
   - `treeVisibilitySlice.ts` — feature-tree visibility toggles for all regions, folders, region folders, and folder selection
 - `helpers/` — pure helpers used by the store
   - `clipping.ts` — clipper-lib wrappers (handles the integer scaling factor): profile↔Clipper-path conversion, boolean/offset execution, and overlap predicates. Arc/curve reconstruction of Clipper output lives in `engine/toolpaths/arcReconstruction.ts`.
-  - `derivedFeatures.ts` — computes derived features from the feature tree
+  - `derivedFeatures.ts` — computes derived features from the feature tree; also createDerivedFeature, previewOffsetFeatures, joinOpenProfiles, and clearStaleConstraints
   - `geometry.ts` — geometric utilities (bounds, transforms)
   - `transform.ts` — point/profile/clamp/tab translation, rotation, mirroring, and affine transforms; arc→bezier conversion
   - `referenceTransforms.ts` — feature/backdrop resize, rotate, mirror from reference geometry; corner fillet radius and application
